Add in a gpu node soft anti-affinity for pods that don't request GPUs
This commit adds in a soft anti-affinity for nodes tagged as GPU
nodes by the Nvidia GPU operator. When creating a pod that doesn't
request any GPUs, it will have an affinity added that avoids GPU
tagged nodes if possible. If there are no available nodes that
satisfy that requirement, it will allow scheduling onto the GPU
nodes to avoid a scheduling failure.